# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Example Blog'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://example-blog:8888'
paths:
...
  /api/password-email:
    post:
      summary: 'Change Password'
      operationId: changePassword
      description: 'Endpoint for sending link to reset password.'
      parameters:
        -
          in: query
          name: email
          description: 'string User email.'
          example: johnsky@example.net
          required: true
          schema:
            type: string
            description: 'string User email.'
            example: johnsky@example.net
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. Must be at least 3 characters. Must not be greater than 75 characters.'
                  example: jweimann@example.net
              required:
                - email
      security: []
...